+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ### Added
 
+- ReqBrPhone and ReqBrMobile classes —
+  [128](https://github.com/dartoos-dev/formdator/issues/128).
 - BrPhone class to validate both masked and unmasked Brazilian landline numbers
   — [108](https://github.com/dartoos-dev/formdator/issues/108).
 - BrMobile class to validate both masked and unmasked Brazilian mobile numbers